Update ImGui to new texture API, add Box, refactor SRV
- Upgrade Hexa.NET.ImGui, ImGui.Widgets, and HexaGen.Runtime packages
- Refactor ImGuiManager/ImGuiRenderer for ImTextureRef/ImTextureID
- Add SRVExtensions.ToTexRef and implicit Icon→ImTextureRef conversion
- Implement ImGui dynamic texture management in ImGuiRenderer
- Add Box struct and UpdateSubresource(Box?) overload for partial updates
- Update ImGuiImplSDL3 usage and support DPI scaling
- Change all ImGui.Image/AddImage calls to use ImTextureRef
- Update ImPlot.PlotScatter to use pointer arguments
- Minor fixes and code cleanup for new ImGui/ImPlot APIs

+++ HexaEngine.Editor/LayoutManager.cs
@@ -120,7 +120,7 @@
         private static unsafe void SetIniString(string value)
         {
             var byteCount = Encoding.UTF8.GetByteCount(value);
-            var str = (byte*)ImGui.MemAlloc((ulong)(byteCount + 1));
+            var str = (byte*)ImGui.MemAlloc((nuint)(byteCount + 1));
             fixed (char* pValue = value)
             {
                 Encoding.UTF8.GetBytes(pValue, value.Length, str, byteCount);
